requirements_efficientdet.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0

